added test case for check_close_game and rearranged app.py into separate folder
